+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add module_options deployment parameter for shared module options
+- Support Terraform backend config via Runway module options
 
 ## [0.42.0] - 2019-03-06
 ### Fixed